Port renamed from gd32vf103v_longan_nano to gd32vf103c_longan_nano as there is an gd32vf103C on this board.
