+++ SoC/gd32vf103/Board/gd32vf103c_longan_nano/Include/gd32vf103c_longan_nano.h
