+++ SoC/gd32vf103/Board/gd32vf103c_longan_nano/Include/nuclei_sdk_hal.h
@@ -6,7 +6,7 @@
  extern "C" {
 #endif
 
-#include "gd32vf103v_longan_nano.h"
+#include "gd32vf103c_longan_nano.h"
 
 #define SOC_DEBUG_UART      GD32_COM0
