+++ SoC/gd32vf103/Board/gd32vf103c_longan_nano/openocd_gd32vf103.cfg
